requirements.txt
click 
pyzmq 

uvicorn
fastapi 
websocket 

pydantic 
pydantic-settings 

async-timeout 
wget 
httpx 